% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{checkRVersion}
\alias{checkRVersion}
\title{Check if your R version is up to date}
\usage{
checkRVersion()
}
\description{
Function to check current R version against the latest release from CRAN
}
