% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_formatting_functions.R
\name{fromExcelDate}
\alias{fromExcelDate}
\title{Function to return an excel date, which can read into R as seconds since 1899-12-30, as a date object}
\usage{
fromExcelDate(date_as_number)
}
\arguments{
\item{date_as_number}{an excel numeric date}
}
\value{
a date object
}
\description{
Function to return an excel date, which can read into R as seconds since 1899-12-30, as a date object
}
