% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_tbl_names}
\alias{get_tbl_names}
\title{Get the names of a BigQuery table}
\usage{
get_tbl_names(tbl, pattern = NULL)
}
\arguments{
\item{tbl}{a BigQuery table}

\item{pattern}{a string or regex pattern used to match the column names in tbl}
}
\value{
a character vector comprised of all column names OR if pattern is specified all column names that match pattern
}
\description{
Function to return a character vector containing all the names of a big query table filtered by pattern if specified
}
