% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{lazy_names}
\alias{lazy_names}
\title{Get the names of a BigQuery table}
\usage{
lazy_names(x)
}
\arguments{
\item{x}{a BigQuery table}
}
\value{
a character vector comprised of all column names
}
\description{
Function to return a character vector containing all the names of a big query table filtered by pattern if specified
}
