% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{list_files}
\alias{list_files}
\title{List files and return the entire local url including parent folders}
\usage{
list_files(dir)
}
\arguments{
\item{dir}{a directory}
}
\value{
a character vector comprised of full paths to all files in the specified directory
}
\description{
Function to list all files in a directory including full paths
}
