% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{pu}
\alias{pu}
\title{Pull a column from a data frame or tibble and return it as a vector of unique values}
\usage{
pu(df, ...)
}
\arguments{
\item{...}{a column}
}
\value{
a character vector comprised of all unique values in a given column in a tibble
}
\description{
Function to pull a column from a data frame or tibble and return it as a vector of unique values
}
