% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{round_half_up}
\alias{round_half_up}
\title{Rounding function to round values of exactly n.5 up}
\usage{
round_half_up(dbl)
}
\arguments{
\item{dbl}{a numeric value}
}
\value{
a rounded numeric value
}
\description{
Function to round values of exactly n.5 up rather than round, which would round even numbers down and odd up:
with round 1.5 and 2.5 would both give 2. round_half_up would return 2 and 3.
}
