% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_formatting_functions.R
\name{row_to_col_names}
\alias{row_to_col_names}
\title{Function to set a given row of a data frame as column headers}
\usage{
row_to_col_names(df, row = 1, clean_col_names = F)
}
\arguments{
\item{df}{a data.frame or tibble}

\item{row}{the row to set as column headers}

\item{clean_col_names}{should the column names be cleaned using janitor::clean_names}

\item{add_index}{should columns be given a trailing unique integer to avoid duplicate names}
}
\value{
a date object
}
\description{
Function to set a given row of a data frame as column headers
}
