% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{source_bq_tbl2}
\alias{source_bq_tbl2}
\title{load a table from the SQL location}
\usage{
source_bq_tbl2(connection, table_location)
}
\arguments{
\item{connection}{a BQ connection}
}
\value{
a lazy query
}
\description{
Function to load a table from the SQL location
}
