% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{sql_copy}
\alias{sql_copy}
\title{Render and copy to clipboard an sql query}
\usage{
sql_copy(query)
}
\arguments{
\item{...}{a column}
}
\value{
a character vector comprised of all unique values in a given column in a tibble
}
\description{
Function to render and copy to clipboard a sql query
}
