% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{str_get_matches}
\alias{str_get_matches}
\title{Function to return all characters in a vector that match a pattern}
\usage{
str_get_matches(c, pattern)
}
\arguments{
\item{c}{a character vector}

\item{pattern}{a string or regex pattern to match}
}
\value{
a character vector comprised of all elements of c that match pattern
}
\description{
Function to return all characters in a vector that match a pattern
}
