% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{tbl_sample_n}
\alias{tbl_sample_n}
\title{Function to return a character vector containing all the names of a big query table, filtered by pattern if specified}
\usage{
tbl_sample_n(tbl, n, distinct_on = NULL, keep_all = T)
}
\arguments{
\item{tbl}{a BigQuery table}

\item{n}{number of ids or rows to sample}

\item{distinct_on}{column on which to call dplyr::distinct. If unspecified function will return n rows.}

\item{keep_all}{should the function find n unique values for distinct_on and return all rows containing those values, T,
or n rows, each with a distinct value for distinct_on.}
}
\value{
Lazy query with with n rows or n unique values of distinct_on
}
\description{
Function to return a character vector containing all the names of a big query table, filtered by pattern if specified
}
